requirements.txt
diffusers==0.35.2
DISTS_pytorch==0.1
lmdb==1.7.5
numpy==2.3.5
Pillow==12.0.0
pytorch_msssim==1.0.0
PyYAML==6.0.3
requests==2.32.5
safetensors==0.7.0
six==1.17.0
scikit-image>=0.24.0
tqdm==4.67.1
transformers==4.57.1
matplotlib==3.8.0